Fix types and factor out ImageUploader
